+++ tests/test_session.py
@@ -5,6 +5,7 @@ import pymongo
 import redis
 import pytest
 import six
+import pymysql
 
 import werobot
 import werobot.testing
@@ -12,6 +13,7 @@ import werobot.utils
 from werobot.session import SessionStorage
 from werobot.session import filestorage, mongodbstorage, redisstorage, saekvstorage
 from werobot.session import sqlitestorage
+from werobot.session import mysqlstorage
 from werobot.utils import to_binary
 
 
@@ -102,7 +104,15 @@ def test_session_storage_delete():
     filestorage.FileStorage(),
     mongodbstorage.MongoDBStorage(pymongo.MongoClient().t.t),
     redisstorage.RedisStorage(redis.Redis()),
-    sqlitestorage.SQLiteStorage()
+    sqlitestorage.SQLiteStorage(),
+    mysqlstorage.MySQLStorage(
+        conn=pymysql.connect(
+            user=os.environ.get('DATABASE_MYSQL_USERNAME', ''),
+            password=os.environ.get('DATABASE_MYSQL_PASSWORD', ''),
+            db='werobot',
+            host='127.0.0.1',
+            charset='utf8'
+        ))
 ])
 def test_storage(storage):
     assert storage.get("喵") == {}